requirements.txt
aiohttp
asyncio
celery
cryptography
django
djangorestframework
flower
gunicorn
psycopg2-binary
pytz